# docs/swagger.json
{
    "openapi": "3.0.0",
    "info": {
      "title": "Uniswap Transaction",
      "description": "A basic flask application to determine the transaction fee based on the transaction hash",
      "version": "1.0"
    },
    "basePath": "/server",
    "schemes": [
      "http"
    ],
    "paths": {
      "/transaction_fee_single": {
        "get": {
          "summary": "Calculates the transaction fee in USDT given a transaction hash",
          "produces": [
            "application/json"
          ],
          "parameters": [{
            "in": "query",
            "name": "hash",
            "required": true,
            "schema": {
                "type": "string"
            },
            "allowedReserved": false
          }],
          "responses": {
            "200": {
              "description": "Successful response",
              "schema": {
                "type": "object",
                "properties": {
                  "message": {
                    "type": "string"
                  }
                }
              }
            },
            "404": {
              "description": "Error response",
              "schema": {
                "type": "object",
                "properties": {
                  "message": {
                    "type": "string"
                  }
                }
              }
            }
          }
        }
      },
      "/transaction_fee_batch": {
        "post": {
          "summary": "Calculates the transaction fee in USDT given a list of transaction hash",
          "consumes": [
            "application/json"
          ],
          "requestBody": {
            "description": "List of transaction hashes",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "hash": {
                      "type": "array",
                        "items": {
                          "type": "string",
                          "example": "0x123123123"
                        },
                        "description": "List of transaction hashes to process"
                    }

                  }
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "Successful response",
              "schema": {
                "type": "object",
                "properties": {
                  "message": {
                    "type": "string"
                  }
                }
              }
            }
          }
        }
      }
    }
  }